Move cache around.
